fix(lint): remove test util
